version 5.1.2
client:
 * fix crash when started on device that has no GPS receiver

server:
 * server changes for cycles ending 2021-09-09, 07-15, 05-20, 03-25

+++ decosects/purgeoldcharts.sh
@@ -70,6 +70,31 @@ function purgedatum
             mv "$file" "purged-$file"
             continue
         fi
+        if [ "${file:0:21}" == "datums/europlatecsvs_" ]
+        then
+            mv "$file" "purged-$file"
+            continue
+        fi
+        if [ "${file:0:21}" == "datums/europlategeos_" ]
+        then
+            mv "$file" "purged-$file"
+            continue
+        fi
+        if [ "${file:0:21}" == "datums/europlategifs_" ]
+        then
+            mv "$file" "purged-$file"
+            continue
+        fi
+        if [ "${file:0:21}" == "datums/europlatelist_" ]
+        then
+            mv "$file" "purged-$file"
+            continue
+        fi
+        if [ "${file:0:21}" == "datums/europlatepdfs_" ]
+        then
+            mv "$file" "purged-$file"
+            continue
+        fi
         if [ "${file:0:12}" == "datums/fixes" ]
         then
             rm -rf "$file"
